+++ src/main.rs
@@ -61,6 +61,7 @@ mod test_utilities;
 mod tile;
 mod tree_library;
 mod version_check;
+mod voxy;
 mod water_depth;
 mod world_editor;
 mod world_utils;